05/requirements.txt
geopandas
matplotlib
streamlit
plotly
folium